TIP-842: Move group type normalizers to structure sub context
