+++ src/Akeneo/Pim/Structure/Component/Normalizer/InternalApi/GroupTypeNormalizer.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Normalizer;
+namespace Akeneo\Pim\Structure\Component\Normalizer\InternalApi;
 
 use Akeneo\Pim\Structure\Component\Model\GroupTypeInterface;
+use Pim\Bundle\EnrichBundle\Normalizer\StructuredVersionProviderInterface;
 use Pim\Bundle\EnrichBundle\Provider\StructureVersion\StructureVersionProviderInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
