+++ src/Akeneo/Pim/Structure/Component/Normalizer/Standard/GroupTypeNormalizer.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Standard;
+namespace Akeneo\Pim\Structure\Component\Normalizer\Standard;
 
 use Akeneo\Pim\Structure\Component\Model\GroupTypeInterface;
+use Pim\Component\Catalog\Normalizer\Standard\TranslationNormalizer;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
