+++ src/Akeneo/Pim/Structure/Component/spec/Normalizer/Standard/GroupTypeNormalizerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Catalog\Normalizer\Standard;
+namespace spec\Akeneo\Pim\Structure\Component\Normalizer\Standard;
 
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Model\GroupTypeInterface;
@@ -15,7 +15,7 @@ class GroupTypeNormalizerSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Pim\Component\Catalog\Normalizer\Standard\GroupTypeNormalizer');
+        $this->shouldHaveType('Akeneo\Pim\Structure\Component\Normalizer\Standard\GroupTypeNormalizer');
     }
 
     function it_is_a_normalizer()
